requirements.txt
invoke==0.21.0
colorlog==2.10.0
python-dotenv==0.7.1
requests==2.18.4